# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Campus Planner Backend",
    "description": "Backend API for AI Campus Planner",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/layout-runs": {
      "post": {
        "tags": [
          "layouts"
        ],
        "summary": "Create Layout Run",
        "operationId": "create_layout_run_api_projects__project_id__layout_runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutRunCreateRequest",
                "default": {
                  "candidateCount": 100,
                  "topK": 5,
                  "algorithm": "GNN_NSGA2"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutRunCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Generation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationFailedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/layouts": {
      "get": {
        "tags": [
          "layouts"
        ],
        "summary": "Get Project Layouts",
        "operationId": "get_project_layouts_api_projects__project_id__layouts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/layouts/{layout_id}": {
      "get": {
        "tags": [
          "layouts"
        ],
        "summary": "Get Layout Detail",
        "operationId": "get_layout_detail_api_layouts__layout_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Layout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/layouts/{layout_id}/select": {
      "post": {
        "tags": [
          "layouts"
        ],
        "summary": "Select Layout",
        "operationId": "select_layout_api_layouts__layout_id__select_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Layout Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutSelectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/layouts/{layout_id}/blueprint": {
      "get": {
        "tags": [
          "layouts"
        ],
        "summary": "Get Blueprint",
        "operationId": "get_blueprint_api_layouts__layout_id__blueprint_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Layout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BlueprintBuildingSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "depth": {
            "type": "number",
            "title": "Depth"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "rotation": {
            "type": "number",
            "title": "Rotation",
            "default": 0.0
          },
          "floorCount": {
            "type": "integer",
            "title": "Floorcount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "zone",
          "x",
          "y",
          "width",
          "depth",
          "height",
          "floorCount"
        ],
        "title": "BlueprintBuildingSchema"
      },
      "BlueprintResponse": {
        "properties": {
          "projectId": {
            "type": "integer",
            "title": "Projectid"
          },
          "layoutId": {
            "type": "integer",
            "title": "Layoutid"
          },
          "site": {
            "$ref": "#/components/schemas/SiteDimensionSchema"
          },
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/BlueprintBuildingSchema"
            },
            "type": "array",
            "title": "Buildings"
          },
          "roads": {
            "items": {},
            "type": "array",
            "title": "Roads"
          },
          "greenAreas": {
            "items": {},
            "type": "array",
            "title": "Greenareas"
          },
          "parkingAreas": {
            "items": {},
            "type": "array",
            "title": "Parkingareas"
          },
          "entrances": {
            "items": {
              "$ref": "#/components/schemas/EntranceSchema"
            },
            "type": "array",
            "title": "Entrances"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "projectId",
          "layoutId",
          "site"
        ],
        "title": "BlueprintResponse"
      },
...
      "CandidateBuildingPosition": {
        "properties": {
          "buildingId": {
            "type": "integer",
            "title": "Buildingid"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "rotation": {
            "type": "number",
            "title": "Rotation",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "buildingId",
          "x",
          "y"
        ],
        "title": "CandidateBuildingPosition"
      },
...
      "EntranceSchema": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "width"
        ],
        "title": "EntranceSchema"
      },
      "LayoutDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "feasible": {
            "type": "boolean",
            "title": "Feasible",
            "default": true
          },
          "site": {
            "$ref": "#/components/schemas/SiteDimensionSchema"
          },
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/CandidateBuildingPosition"
            },
            "type": "array",
            "title": "Buildings"
          },
          "roads": {
            "items": {},
            "type": "array",
            "title": "Roads"
          },
          "greenAreas": {
            "items": {},
            "type": "array",
            "title": "Greenareas"
          },
          "parkingAreas": {
            "items": {},
            "type": "array",
            "title": "Parkingareas"
          },
          "entrances": {
            "items": {
              "$ref": "#/components/schemas/EntranceSchema"
            },
            "type": "array",
            "title": "Entrances"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "rank",
          "site"
        ],
        "title": "LayoutDetailResponse"
      },
      "LayoutListResponse": {
        "properties": {
          "layouts": {
            "items": {
              "$ref": "#/components/schemas/LayoutSummarySchema"
            },
            "type": "array",
            "title": "Layouts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "layouts"
        ],
        "title": "LayoutListResponse"
      },
      "LayoutMetricsSchema": {
        "properties": {
          "landUtilization": {
            "type": "number",
            "title": "Landutilization"
          },
          "greenRatio": {
            "type": "number",
            "title": "Greenratio"
          },
          "parkingRatio": {
            "type": "number",
            "title": "Parkingratio",
            "default": 0.0
          },
          "accessibilityScore": {
            "type": "number",
            "title": "Accessibilityscore",
            "default": 0.0
          },
          "roadEfficiency": {
            "type": "number",
            "title": "Roadefficiency",
            "default": 0.0
          },
          "constraintScore": {
            "type": "number",
            "title": "Constraintscore",
            "default": 1.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "landUtilization",
          "greenRatio"
        ],
        "title": "LayoutMetricsSchema"
      },
      "LayoutRunCreateRequest": {
        "properties": {
          "candidateCount": {
            "type": "integer",
            "title": "Candidatecount",
            "default": 100
          },
          "topK": {
            "type": "integer",
            "title": "Topk",
            "default": 5
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm",
            "default": "GNN_NSGA2"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LayoutRunCreateRequest"
      },
      "LayoutRunCreateResponse": {
        "properties": {
          "runId": {
            "type": "integer",
            "title": "Runid"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "COMPLETED"
          },
          "layoutCount": {
            "type": "integer",
            "title": "Layoutcount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runId",
          "layoutCount"
        ],
        "title": "LayoutRunCreateResponse"
      },
      "LayoutSelectResponse": {
        "properties": {
          "projectId": {
            "type": "integer",
            "title": "Projectid"
          },
          "layoutId": {
            "type": "integer",
            "title": "Layoutid"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "SELECTED"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "projectId",
          "layoutId"
        ],
        "title": "LayoutSelectResponse"
      },
      "LayoutSummarySchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "feasible": {
            "type": "boolean",
            "title": "Feasible",
            "default": true
          },
          "metrics": {
            "$ref": "#/components/schemas/LayoutMetricsSchema"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "rank",
          "metrics"
        ],
        "title": "LayoutSummarySchema"
      },
...
      "SiteDimensionSchema": {
        "properties": {
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "width",
          "height"
        ],
        "title": "SiteDimensionSchema"
      },
...
      "ValidationErrorResponse": {
        "title": "ValidationErrorResponse",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string",
            "example": "VALIDATION_ERROR"
          }
        },
        "required": [
          "detail"
        ]
      },
      "GenerationFailedResponse": {
        "title": "GenerationFailedResponse",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string",
            "example": "GENERATION_FAILED"
          }
        },
        "required": [
          "detail"
        ]
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}